+++ documentation/changelog.md
@@ -56,6 +56,7 @@ description: List of all the changes and current versions.
 + Fix Length Error for Ped Spawning
 + Some Interface Work
 + Fix Marker Scale for Jobs
++ Make Creation Interfaces Return UID(s)
 ```
 
 ## 2.0.3